% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mem.usage}
\alias{mem.usage}
\title{Returns the memory usage in KB of a process with the specified process id. By
default, returns the memory usage of the current R process.}
\usage{
mem.usage(pid = getpid(), remote = "")
}
\arguments{
\item{pid}{Process ID (default is the current process id).}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass an empty string "" (default).}
}
\description{
Returns the memory usage in KB of a process with the specified process id. By
default, returns the memory usage of the current R process.
}

