% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ssh.utils}
\alias{ssh.utils}
\alias{ssh.utils-package}
\title{Utility Functions to Run Remote and Local System Commands.}
\description{
Package \code{ssh.utils} provides utility functions for calling system
commands and capturing their output, both locally and remotely using ssh/scp.
}
\section{Maintainer}{
 Sergei Izrailev
}

\section{Copyright}{
 Copyright (C) Collective, Inc.
}

\section{License}{
 Apache License, Version 2.0,
   available at http://www.apache.org/licenses/LICENSE-2.0
}

\section{URL}{
 http://github.com/collectivemedia/ssh.utils
}

\section{Installation from github}{

\code{devtools::install_github("collectivemedia/ssh.utils")}
}
\author{
Sergei Izrailev
}
\seealso{
\code{\link{run.remote}}, \code{\link{cp.remote}}
}
\keyword{"capture}
\keyword{"system}
\keyword{command"}
\keyword{output"}
\keyword{remote}
\keyword{scp}
\keyword{ssh}

