% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ssh.utils}
\alias{ssh.utils}
\alias{ssh.utils-package}
\title{Local and remote system commands with output and error capture.}
\description{
Package \code{ssh.utils} provides utility functions for system command
execution, both locally and remotely using ssh/scp. The command
output is captured and provided to the caller. This functionality is
intended to streamline calling shell commands from R, retrieving and
using their output, while instrumenting the calls with appropriate
error handling.
}
\section{OS_type}{
 unix
}

\section{Maintainer}{
 Sergei Izrailev
}

\section{Copyright}{
 Copyright (C) Collective, Inc.
}

\section{License}{
 Apache License, Version 2.0,
   available at http://www.apache.org/licenses/LICENSE-2.0
}

\section{URL}{
 http://github.com/collectivemedia/ssh.utils
}

\section{Installation from github}{

\code{devtools::install_github("collectivemedia/ssh.utils")}
}
\author{
Sergei Izrailev
}
\seealso{
\code{\link{run.remote}}, \code{\link{cp.remote}},
         \code{\link{file.exists.remote}}, \code{\link{mkdir.remote}},
         \code{\link{ps.grep.remote}}, \code{\link{mem.usage}}
}
\keyword{"capture}
\keyword{"system}
\keyword{bash}
\keyword{command"}
\keyword{output"}
\keyword{remote}
\keyword{scp}
\keyword{shell}
\keyword{ssh}

